fix: delete stories

+++ functions/index.js
@@ -6,14 +6,16 @@ admin.initializeApp();
 exports.deleteStories24Hours = functions.pubsub.schedule("every 12 hours").onRun(async(context)=>{
     const now = Date.now();
     const oneDay = 24 * 60 * 60 * 1000;
-    const yesterday = new Date(now - oneDay);
+    const yesterday = now - oneDay;
     await admin.firestore().collection('/stories').where('createdAt', '<', yesterday)
-    .get().then( (snapshot)=> {
-        snapshot.forEach((doc)=> {
-             doc.ref.delete()
-             .then((res) => console.log(res))
-             .catch((err) => console.log(err))
-        });
+    .get().then((snapshot)=> {
+        if (!snapshot.empty){
+            snapshot.forEach((doc)=> {
+                doc.ref.delete()
+                .then((res) => console.log(res))
+                .catch((err) => console.log(err))
+            });
+        }
         return ;
     })
     .catch((err)=>{
